requirements.txt
yagmail-0.14.260